Update ALetter.py
Assigns directly to the vars, instead of appending, since its always initialized to 0/empty string.

The check for grabbing the first letter of the serial was excluding the letter "Z", since the ascii for "Z" is 90. It now checks correctly.

Since the var "alphabet' is already a list, there is no need to convert it to a list again, it now copies the elements in the list.

User input is now case insensitive.